pull out schema functions
